fix: Add Psion website dump from the Wayback Machine (#138)
Thanks to @scienceapps, there’s now an archive of many zips from old
sites on the Wayback Machine. This change adds that in advance of more
comprehensive support for Wayback Machine archives.

This change also adds the beginnings of a management CLI tool that will
hopefully make it quicker/easier to add sources in the future.